move mobile and desktop view together in one file
